Includes custom "DeepRequired" type in the build
