+++ src/utils/handleRequest.ts
@@ -3,6 +3,7 @@ import { MockedRequest, RequestHandler } from '../handlers/RequestHandler'
 import { ServerLifecycleEventsMap } from '../node/glossary'
 import { MockedResponse } from '../response'
 import { SharedOptions } from '../sharedOptions'
+import { DeepRequired } from '../typeUtils'
 import { ResponseLookupResult, getResponse } from './getResponse'
 import { onUnhandledRequest } from './request/onUnhandledRequest'
 import { readResponseCookies } from './request/readResponseCookies'
